J48 loading dataset & generating 2 values

+++ src/main/java/cz/uhk/fim/project/bakalarka/model/Run.java
@@ -47,7 +47,7 @@ public class Run {
         this.id = id;
     }
 
-    public double getLenglhinmeters() {
+    public Integer getLenglhinmeters() {
         return lenglhinmeters;
     }
