+++ src/main/java/cz/uhk/fim/project/bakalarka/model/Swimming.java
@@ -47,7 +47,7 @@ public class Swimming {
         this.id = id;
     }
 
-    public double getLenglhinmeters() {
+    public Integer getLenglhinmeters() {
         return lenglhinmeters;
     }
